In this PR, error messages thrown from backend for student/course/section not found are made more user-friendly with instructions on how to fix the error.
Added checkbox in modal to create a new  course/section from the csv file if it does not exist in the database already. If the checkbox is not ticked, usual errors are displayed.

To test, upload csv file (first ensure student with studentNum exists in db), verify errors are displayed correctly if checkbox is not ticked. Next, tick the checkbox and import should succeed with new course/section added.